requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.9
Pillow==7.1.2
PySide2==5.15.0
requests==2.24.0
shiboken2==5.15.0
urllib3==1.25.9
